+++ src/services/files.service.ts
@@ -1,10 +1,16 @@
-import { getStorage } from "firebase-admin/storage";
+import { getStorage, Storage } from "firebase-admin/storage";
 import { Request, Response } from "express";
 import AppResult from "./../errors/app-error";
-import { AppErrorsMessages, AppSuccessMessages } from "../constants";
+import {
+  AppErrorsMessages,
+  AppSuccessMessages,
+  STORAGE_BUCKETS,
+} from "../constants";
 import moment from "moment";
 import { getUserById } from "../services/user.service";
 import log from "../utils/logs";
+import { IImageDetails } from "../models/files.models";
+import { getImageThumbnail } from "../utils";
 
 export const uploadFileToStorage = async (req: Request, res: Response) => {
   const image: any = req.file;
@@ -18,9 +24,7 @@ export const uploadFileToStorage = async (req: Request, res: Response) => {
       .json(new AppResult(AppErrorsMessages.USER_NOT_FOUND, null, 400));
   }
 
-  const bucket = process.env.FIREBASE_STORAGE_BUCKET_URL;
-
-  if (!bucket) {
+  if (!STORAGE_BUCKETS.socialbioProject) {
     return res
       .status(500)
       .json(
@@ -62,7 +66,7 @@ export const uploadFileToStorage = async (req: Request, res: Response) => {
     log.success("File upload successful.");
     (
       req as any
-    ).file.firebaseUrl = `https://storage.googleapis.com/${bucket}/${fileName}`;
+    ).file.firebaseUrl = `${STORAGE_BUCKETS.baseUrl}/${STORAGE_BUCKETS.socialbioProject}/${fileName}`;
 
     await fileToSave.makePublic();
     return res.status(200).json(image.firebaseUrl);
@@ -82,9 +86,7 @@ export const deleteFileFromStorage = async (req: Request, res: Response) => {
       .json(new AppResult(AppErrorsMessages.USER_NOT_FOUND, null, 400));
   }
 
-  const bucket = process.env.FIREBASE_STORAGE_BUCKET_URL;
-
-  if (!bucket) {
+  if (!STORAGE_BUCKETS.socialbioProject) {
     return res
       .status(500)
       .json(
@@ -98,7 +100,10 @@ export const deleteFileFromStorage = async (req: Request, res: Response) => {
 
   const storage = getStorage().bucket();
 
-  const fileName = url.replace(`https://storage.googleapis.com/${bucket}/`, "");
+  const fileName = url.replace(
+    `https://storage.googleapis.com/${STORAGE_BUCKETS.socialbioProject}/`,
+    ""
+  );
 
   return storage
     .file(fileName)
@@ -119,13 +124,14 @@ export const deleteFileFromStorage = async (req: Request, res: Response) => {
 
 // TODO: Check if can be used into `deleteFileFromStorage()`
 export const deleteFile = async (url: string) => {
-  const bucket = process.env.FIREBASE_STORAGE_BUCKET_URL;
-
-  if (!bucket) return false;
+  if (!STORAGE_BUCKETS.socialbioProject) return false;
 
   const storage = getStorage().bucket();
 
-  const fileName = url.replace(`https://storage.googleapis.com/${bucket}/`, "");
+  const fileName = url.replace(
+    `${STORAGE_BUCKETS.baseUrl}/${STORAGE_BUCKETS.socialbioProject}/`,
+    ""
+  );
 
   return storage
     .file(fileName)
@@ -135,9 +141,7 @@ export const deleteFile = async (url: string) => {
 };
 
 export const deleteAllUserFiles = async (userId: string): Promise<number> => {
-  const bucket = process.env.FIREBASE_STORAGE_BUCKET_URL;
-
-  if (!bucket) return 0;
+  if (!STORAGE_BUCKETS.socialbioProject) return 0;
   const userBucketName = `users/${userId}`;
 
   const userStorage = getStorage().bucket(userBucketName);
@@ -149,4 +153,85 @@ export const deleteAllUserFiles = async (userId: string): Promise<number> => {
     .delete()
     .then(() => filesCount)
     .catch(() => 0);
+};
+
+export const getAllImagesOnBucket = async (
+  bucket: string
+): Promise<IImageDetails[] | null> => {
+  if (!STORAGE_BUCKETS.socialbioProject) return null;
+
+  const [files] = await getStorage().bucket().getFiles({
+    prefix: bucket,
+    delimiter: "/",
+  });
+
+  const urlArray: IImageDetails[] = [];
+
+  files.forEach((file: any) => {
+    if (
+      file &&
+      file.name &&
+      (file.name as string).length > 0 &&
+      file.name.length > bucket.length // Preventing an case that GCloud returned an empty file
+    ) {
+      let originalUrl = "";
+      let thumbnailUrl = "";
+
+      if (!(file.name as string).includes("_200x200")) {
+        originalUrl = `${STORAGE_BUCKETS.baseUrl}/${STORAGE_BUCKETS.socialbioProject}/${file.name}`;
+        thumbnailUrl = getImageThumbnail(originalUrl, 200);
+      }
+
+      if (originalUrl && thumbnailUrl) {
+        urlArray.push({
+          original: originalUrl,
+          thumbnail: thumbnailUrl,
+        });
+      }
+    }
+  });
+
+  if (urlArray && urlArray.length > 0) return urlArray;
+
+  return null;
+};
+
+export const getAllUserImages = async (
+  userId: string
+): Promise<IImageDetails[] | null> => {
+  const userBucket = `users/${userId}/${STORAGE_BUCKETS.userUploadedImages}/`;
+
+  return await getAllImagesOnBucket(userBucket);
+};
+
+export const getAllButtonsTemplates = async (): Promise<
+  IImageDetails[] | null
+> => {
+  const buttonsTemplatesBucket = `${STORAGE_BUCKETS.buttonsTemplates}/`;
+
+  return await getAllImagesOnBucket(buttonsTemplatesBucket);
+};
+
+export const getAllBackgroundsTemplates = async (): Promise<
+  IImageDetails[] | null
+> => {
+  const bgTemplatesBucket = `${STORAGE_BUCKETS.backgroundsTemplates}/`;
+
+  return await getAllImagesOnBucket(bgTemplatesBucket);
+};
+
+export const getAllUserProfileTemplates = async (): Promise<
+  IImageDetails[] | null
+> => {
+  const userProfileTemplatesBucket = `${STORAGE_BUCKETS.userProfileTemplates}/`;
+
+  return await getAllImagesOnBucket(userProfileTemplatesBucket);
+};
+
+export const getAllPagesImgsTemplates = async (): Promise<
+  IImageDetails[] | null
+> => {
+  const pagesImgsTemplatesBucket = `${STORAGE_BUCKETS.pagesImgsTemplates}/`;
+
+  return await getAllImagesOnBucket(pagesImgsTemplatesBucket);
 };