+++ src/utils/index.ts
@@ -6,3 +6,14 @@ export const replaceAllSpacesByUnderlines = (
   if (lowerCase) result.toLowerCase();
   return result;
 };
+
+export const getImageThumbnail = (
+  imagePathAndName: string,
+  thumbSize: number
+) => {
+  const splitPath = imagePathAndName.split("/");
+  const [fileName, fileExtension] = splitPath[splitPath.length - 1].split(".");
+  return `${splitPath
+    .slice(0, splitPath.length - 1)
+    .join("/")}/${fileName}_${thumbSize}x${thumbSize}.${fileExtension}`;
+};