+++ changelog.md
@@ -16,6 +16,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Versions
 
+### [1.6.0] - 2022-10-03
+
+- Feature: Implemented endpoints to get:
+  - All user images
+  - All buttons templates
+  - All backgrounds templates
+  - All pages images templates
+
 ### [1.5.1] - 2022-10-03
 
 - Fix: Deleteing files when user deletes the account